refactor(registry)!: remove core ai module and consolidate transform helpers
Delete tracecat_registry.core.ai and remove core.ai.extract registry fixture wiring.

Move flatten helpers into expression functions, add drop_nulls/flatten_dict/flatten_json to FN mapping, and shift flatten_json coverage to unit function tests.

Remove legacy core.transform actions/tests (is_in, not_in, compact, flatten_json, eval_jsonpaths) and rename lambda action titles for clarity.

BREAKING CHANGE: tracecat_registry.core.ai is removed, and legacy core.transform actions (is_in, not_in, compact, flatten_json, eval_jsonpaths) are no longer available.
